Added two methods to Molecule class. The class now works in controller
