+++ Group22Project/src/application/Valid_Atoms_Test.java
@@ -5,7 +5,7 @@ public class Valid_Atoms_Test {
 	public static void main(String[] args) {
 		
 		
-		Valid_Atoms test_me = new Valid_Atoms();
+		Atom test_me = new Atom();
 		
 		
 		//I have three methods at the moment, getElement, getSymbol and Test
